Getting started

LightGraphsMatching.jl: matching algorithms for LightGraphs

LightGraphsMatching.MatchingResult — Type.

type MatchingResult{T}
    weight::T
    mate::Vector{Int}
end

A type representing the result of a matching algorithm.

weight: total weight of the matching

mate:    `mate[i] = j` if vertex `i` is matched to vertex `j`.
         `mate[i] = -1` for unmatched vertices.

source

LightGraphsMatching.maximum_weight_matching — Function.

maximum_weight_matching{T <:Real}(g::Graph, w::Dict{Edge,T} = Dict{Edge,Int64}())

Given a graph g and an edgemap w containing weights associated to edges, returns a matching with the maximum total weight. w is a dictionary that maps edges i => j to weights. If no weight parameter is given, all edges will be considered to have weight 1 (results in max cardinality matching)

The efficiency of the algorithm depends on the input graph:

If the graph is bipartite, then the LP relaxation is integral.

If the graph is not bipartite, then it requires a MIP solver and

the computation time may grow exponentially.

The package JuMP.jl and one of its supported solvers is required.

Returns MatchingResult containing:

a solve status (indicating whether the problem was solved to optimality)

the optimal cost

a list of each vertex's match (or -1 for unmatched vertices)

source

LightGraphsMatching.maximum_weight_maximal_matching — Function.

maximum_weight_maximal_matching{T<:Real}(g, w::Dict{Edge,T})
maximum_weight_maximal_matching{T<:Real}(g, w::Dict{Edge,T}, cutoff)

Given a bipartite graph g and an edgemap w containing weights associated to edges, returns a matching with the maximum total weight among the ones containing the greatest number of edges.

Edges in g not present in w will not be considered for the matching.

The algorithm relies on a linear relaxation on of the matching problem, which is guaranteed to have integer solution on bipartite graps.

Eventually a cutoff argument can be given, to reduce computational times excluding edges with weights lower than the cutoff.

The package JuMP.jl and one of its supported solvers is required.

The returned object is of type MatchingResult.

source

LightGraphsMatching.minimum_weight_perfect_matching — Function.

minimum_weight_perfect_matching{T<:Real}(g, w::Dict{Edge,T})
minimum_weight_perfect_matching{T<:Real}(g, w::Dict{Edge,T}, cutoff)

Given a graph g and an edgemap w containing weights associated to edges, returns a matching with the mimimum total weight among the ones containing exactly nv(g)/2 edges.

Edges in g not present in w will not be considered for the matching.

This function relies on the BlossomV.jl package, a julia wrapper around Kolmogorov's BlossomV algorithm.

Eventually a cutoff argument can be given, to the reduce computational time excluding edges with weights higher than the cutoff.

The returned object is of type MatchingResult.

In case of error try to change the optional argument tmaxscale (default is tmaxscale=10).

source

LightGraphsMatching.dict_to_arr — Method.

Returns an array of mates from a dictionary that maps edges to {0,1}

source

LightGraphsMatching.cutoff_weights — Method.

cutoff_weights copies the weight matrix with all elements below cutoff set to 0

source

This is the documentation page for LightGraphsMatching. In all documentation examples, we assume LightGraphsMatching has been imported into scope and that LightGraphs is available with the alias lg:

using LightGraphsMatching
import LightGraphs
const lg = LightGraphs